moving WrongConfigVersionException to libs/configuration/configuration-core
