+++ components/configuration/configuration-rpcops-service-impl/build.gradle
@@ -12,7 +12,6 @@ dependencies {
     implementation platform("net.corda:corda-api:$cordaApiVersion")
 
     implementation project(':components:configuration:configuration-read-service')
-    implementation project(':components:configuration:configuration-write-service')
     implementation project(':libs:configuration:configuration-core')
     implementation project(':libs:configuration:configuration-endpoints')
     implementation project(':libs:configuration:configuration-validation')
