+++ components/configuration/configuration-write-service-impl/src/main/kotlin/net/corda/configuration/write/impl/ConfigWriteEventHandler.kt
@@ -1,10 +1,10 @@
 package net.corda.configuration.write.impl
 
-import net.corda.configuration.write.ConfigWriteServiceException
 import net.corda.configuration.write.impl.writer.ConfigurationManagementRPCSubscription
 import net.corda.configuration.write.impl.writer.RPCSubscriptionFactory
 import net.corda.configuration.write.publish.ConfigPublishService
 import net.corda.db.connection.manager.DbConnectionManager
+import net.corda.libs.configuration.exception.ConfigWriteServiceException
 import net.corda.libs.configuration.SmartConfig
 import net.corda.libs.configuration.merger.ConfigMerger
 import net.corda.lifecycle.LifecycleCoordinator
