+++ libs/configuration/configuration-core/src/main/kotlin/net/corda/libs/configuration/exception/ConfigWriteServiceException.kt
@@ -1,4 +1,4 @@
-package net.corda.configuration.write
+package net.corda.libs.configuration.exception
 
 import net.corda.v5.base.exceptions.CordaRuntimeException
 
@@ -6,4 +6,4 @@ import net.corda.v5.base.exceptions.CordaRuntimeException
 open class ConfigWriteServiceException(message: String, e: Exception? = null) : CordaRuntimeException(message, e)
 
 /** Indicates that a configuration management request's version did not match the current version. */
-class ConfigVersionConflictException(message: String, e: Exception? = null): ConfigWriteServiceException(message, e)
+class WrongConfigVersionException(message: String, e: Exception? = null): ConfigWriteServiceException(message, e)
